Stricter type checking.
We need to provide a type for the return value of `load` and
`load_cache`, otherwise mypy is unable to infer the types of calling any
methods on an instance returned from there, defeating the whole point of
annotating the class. Since the class is not yet defined at that point,
we use a string literal to define the type. (Once we drop support for
Python 3.6, we can use `from __future__ import annotations` which will
allow us to use the class itself there without breaking the code.)

For completeness, also annotate Mutex by making it a generic and
declaring a concrete type for the calendar _cache classvar.